refactor: 도메인이 인프라를 의존하던 부분 수정.
- CardShuffler 인터페이스를 도메인 계층, 구현체를 인프라 계층으로 이동
