+++ src/main/java/domain/CardShuffler.java
@@ -1,4 +1,4 @@
-package infra;
+package domain;
 
 public interface CardShuffler {
     int getRandomCardIndex(int deckSize);
